Tests moved to avoid module import errors
